+++ pages/my/infoFix.tsx
@@ -15,8 +15,9 @@ export default function myInfoFix() {
   const [myfixedInfo, setMyfixedInfo] = useState(myInfo?.data.data);
   const [openModal, setOpenModal] = useState(false);
 
+  const { mutate: handleHandInfo } = usePutmyInfo(myfixedInfo);
+
   const handleMoveToMyInfo = () => {
-    console.log('최종 제출', myfixedInfo);
     if (
       myfixedInfo?.shippingInfo.receiverName !== '' &&
       myfixedInfo?.phoneNumber !== '' &&
@@ -27,9 +28,8 @@ export default function myInfoFix() {
       myfixedInfo?.bankName !== '' &&
       myfixedInfo?.accountNumber !== ''
     ) {
-      // alert('제출완료');
-      // Router.push('/my/info');
-      // usePutmyInfo(myfixedInfo);
+      Router.push('/my/info');
+      handleHandInfo();
     } else {
       setOpenModal(true);
     }
@@ -42,10 +42,6 @@ export default function myInfoFix() {
   useEffect(() => {
     setMyfixedInfo(myInfo?.data.data);
   }, [myInfo]);
-
-  useEffect(() => {
-    console.log('변경', myfixedInfo);
-  }, [myfixedInfo]);
 
   if (isLoading) return <div>로딩중...</div>;
   if (error) return <div>에러발생</div>;
